requirements.txt
PyPDF2
pytesseract
pdf2image
spacy
sentence-transformers
pinecone-client
openai
